voucher_client: DRY check+verify
